+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 3.0.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Documento now set b_size=200 by default to show more than 25 items when getting its data.
+  [cekk]
 
 
 3.0.1 (2021-05-04)